rename httpsUrlInput.js -> HttpsUrlInput.jsx
